requirements.txt
click
pytest-runner
fuzzywuzzy
